+++ cli/banner.go
@@ -1,3 +1,3 @@
 package cli
 
-const versions = "v1.0.8"
+const versions = "v1.0.9"