+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ## features
 - Add dangerous image used checking in Docker
 - Add Docker Swarm Service checking
+- Add checking of ephemeral-storage usage
+
+## improvements
+- Annotate the tag of image checking
 
 ## improvements
 - Add unauthorized kubelet checking for each node