+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.9 (2023.8.8)
+## improvements
+- Add the date of kernel compiling checking in checking of kernel version
+
 # 1.0.8 (2023.6.6)
 ## features
 - Add dangerous image used checking in Docker